sharding id 类型由byte改为int

+++ c-delay-queue/src/main/java/com/cjyfff/dq/task/mapper/DelayTaskMapper.java
@@ -21,7 +21,7 @@ public interface DelayTaskMapper {
     int updateStatusByTaskIdAndOldStatus(@Param(value = "taskId") String taskId,
                                          @Param(value = "oldStatus") Integer oldStatus,
                                          @Param(value = "newStatus") Integer newStatus,
-                                         @Param(value = "shardingId") Byte shardingId);
+                                         @Param(value = "shardingId") Integer shardingId);
 
     DelayTask selectByTaskId(@Param(value = "taskId") String taskId);
 
@@ -29,17 +29,17 @@ public interface DelayTaskMapper {
 
     DelayTask selectByTaskIdAndStatus(@Param(value = "oldStatus") Integer oldStatus,
                                       @Param(value = "taskId") String taskId,
-                                      @Param(value = "shardingId") Byte shardingId);
+                                      @Param(value = "shardingId") Integer shardingId);
 
     List<DelayTask> selectByStatusAndExecuteTime(@Param(value = "oldStatus") Integer oldStatus,
-                                                 @Param(value = "shardingId") Byte shardingId,
+                                                 @Param(value = "shardingId") Integer shardingId,
                                                  @Param(value = "executeTimeBegin") Long executeTimeBegin,
                                                  @Param(value = "executeTimeEnd") Long executeTimeEnd);
 
     List<DelayTask> selectByStatusForUpdate(@Param(value = "status") Integer status);
 
     List<DelayTask> selectByStatusAndShardingId(@Param(value = "status") Integer status,
-                                                         @Param(value = "shardingId") Byte shardingId);
+                                                         @Param(value = "shardingId") Integer shardingId);
 
     int updateByPrimaryKey(DelayTask record);
 }
