+++ c-delay-queue/src/main/java/com/cjyfff/dq/task/transport/TestController.java
@@ -52,7 +52,7 @@ public class TestController {
 @Getter
 @Setter
 class TestInfoReqVo {
-    private Byte shardingId;
+    private Integer shardingId;
 
     private String taskId;
 }
