+++ c-delay-queue/src/main/java/com/cjyfff/dq/task/transport/protocol/Packet.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 public abstract class Packet {
     private Byte type;
 
-    private Byte shardingId;
+    private Integer shardingId;
 }
